# Assignment 4/openapi.yaml
openapi: 3.0.3

info:
  title: CampusEats Order Service API
  version: 1.0.0
  description: REST API for managing CampusEats orders.

servers:
  - url: http://localhost:5000

paths:
  /orders:
    post:
      summary: Create an order
      operationId: createOrder
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
      responses:
        '201':
          description: Order created
          headers:
            Location:
              description: URL of the newly created order
              schema:
                type: string
                example: /orders/ORD-1001
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'

    get:
      summary: List orders
      operationId: listOrders
      parameters:
        - name: status
          in: query
          required: false
          description: Filter orders by status.
          schema:
            type: string
            example: PLACED
      responses:
        '200':
          description: List of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'
        '400':
          $ref: '#/components/responses/BadRequest'

    options:
      summary: Describe supported methods for orders
      operationId: ordersOptions
      responses:
        '204':
          description: No content. The Allow header lists supported methods.
          headers:
            Allow:
              description: Methods supported by the /orders resource.
              schema:
                type: string
                example: GET, POST, OPTIONS

  /orders/{orderId}:
    get:
      summary: Get an order
      operationId: getOrder
      parameters:
        - $ref: '#/components/parameters/OrderId'
      responses:
        '200':
          description: Order details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '404':
          $ref: '#/components/responses/NotFound'

  /orders/{orderId}/cancellation:
    post:
      summary: Cancel an order
      operationId: cancelOrder
      parameters:
        - $ref: '#/components/parameters/OrderId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CancellationRequest'
      responses:
        '200':
          description: Order cancelled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CancellationResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'

components:
  parameters:
    OrderId:
      name: orderId
      in: path
      required: true
      description: Unique order identifier.
      schema:
        type: string
        example: ORD-1001

    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: Key used to safely retry order creation without creating a duplicate order.
      schema:
        type: string
        example: 8f4d1c2a-1234-4567-8901-abcdef123456

  schemas:
    CreateOrderRequest:
      type: object
      required:
        - userId
        - cartId
        - deliveryAddressId
        - paymentMethod
      properties:
        userId:
          type: string
          example: CE-STU-33217
        cartId:
          type: string
          example: CART-1001
        deliveryAddressId:
          type: string
          example: ADDR-204
        paymentMethod:
          type: string
          enum:
            - UPI
            - CARD
            - WALLET
          example: UPI
        couponCode:
          type: string
          nullable: true
          example: SAVE10

    Order:
      type: object
      required:
        - orderId
        - status
        - totalAmount
        - estimatedDeliveryMinutes
      properties:
        orderId:
          type: string
          example: ORD-1001
        status:
          type: string
          example: PLACED
        totalAmount:
          type: number
          format: double
          example: 349.00
        estimatedDeliveryMinutes:
          type: integer
          example: 25

    CancellationRequest:
      type: object
      required:
        - reason
      properties:
        reason:
          type: string
          example: Changed my mind

    CancellationResponse:
      type: object
      required:
        - orderId
        - status
        - refundInitiated
      properties:
        orderId:
          type: string
          example: ORD-1001
        status:
          type: string
          example: CANCELLED
        refundInitiated:
          type: boolean
          example: true

    Problem:
      type: object
      required:
        - type
        - title
        - status
        - detail
      properties:
        type:
          type: string
          example: https://campuseats.example/problems/order-not-found
        title:
          type: string
          example: Order not found
        status:
          type: integer
          example: 404
        detail:
          type: string
          example: No order exists with the supplied orderId.

  responses:
    BadRequest:
      description: Malformed request or invalid query parameter.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'

    NotFound:
      description: Requested resource does not exist.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'

    Conflict:
      description: Request conflicts with the current state of the resource.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'

    UnprocessableEntity:
      description: Request is valid but refused by the domain.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'